Handle register address error
