+++ client/src/utils.ts
@@ -1,13 +1,12 @@
 import { EdgeAccount, EdgeCurrencyWallet } from 'edge-core-js';
 import isEmpty from 'lodash/isEmpty';
-import apis from './api/index';
+
 import {
   CartItem,
   DeleteCartItem,
   Domain,
   Prices,
   FioNameItemProps,
-  RegistrationResult,
   WalletKeysObj,
 } from './types';
 
@@ -270,122 +269,6 @@ export const isDomain = (fioName: string) =>
   fioName.indexOf(FIO_ADDRESS_DELIMITER) < 0;
 export const hasFioAddressDelimiter = (value: string): boolean =>
   value.indexOf(FIO_ADDRESS_DELIMITER) > 0;
-
-export const transformResult = ({
-  result,
-  cart,
-  prices,
-}: {
-  result: RegistrationResult;
-  cart: CartItem[];
-  prices: Prices;
-}) => {
-  const errItems = [];
-  const regItems = [];
-
-  const { registered, errors, partial } = result;
-
-  const updatedCart = [...cart];
-
-  const {
-    fio: { address: addressCostFio, domain: domainCostFio },
-    usdt: { address: addressCostUsdc, domain: domainCostUsdc },
-  } = prices;
-
-  if (!isEmpty(errors)) {
-    for (const item of errors) {
-      const { fioName, error, isFree, cartItemId } = item;
-
-      const retObj: CartItem = {
-        id: fioName,
-        domain: '',
-      };
-
-      const partialIndex = partial && partial.indexOf(cartItemId);
-      if (!isDomain(fioName)) {
-        const name = fioName.split('@');
-        const addressName = name[0];
-        const domainName = name[1];
-
-        retObj.address = addressName;
-        retObj.domain = domainName;
-        retObj.error = error;
-
-        if (isFree) {
-          retObj.isFree = isFree;
-        } else {
-          if (
-            cart.find(
-              cartItem =>
-                cartItem.id === cartItemId && cartItem.hasCustomDomain,
-            ) &&
-            partialIndex < 0
-          ) {
-            retObj.costFio = addressCostFio + domainCostFio;
-            retObj.costUsdc = addressCostUsdc + domainCostUsdc;
-          } else {
-            retObj.costFio = addressCostFio;
-            retObj.costUsdc = addressCostUsdc;
-          }
-        }
-      } else {
-        retObj.domain = fioName;
-        retObj.costFio = domainCostFio;
-        retObj.costUsdc = domainCostUsdc;
-      }
-
-      errItems.push(retObj);
-      if (partialIndex > 0) {
-        updatedCart.splice(partialIndex, 1, retObj);
-      }
-    }
-  }
-
-  if (!isEmpty(registered)) {
-    for (const item of registered) {
-      const { fioName, isFree, fee_collected } = item;
-
-      const retObj: CartItem = {
-        id: fioName,
-        domain: '',
-      };
-
-      if (!isDomain(fioName)) {
-        const name = fioName.split('@');
-        const addressName = name[0];
-        const domainName = name[1];
-
-        retObj.address = addressName;
-        retObj.domain = domainName;
-
-        if (isFree) {
-          retObj.isFree = isFree;
-        } else {
-          retObj.costFio = apis.fio.sufToAmount(fee_collected);
-          retObj.costUsdc =
-            (apis.fio.sufToAmount(fee_collected) * addressCostUsdc) /
-            addressCostFio;
-        }
-      } else {
-        retObj.domain = fioName;
-        retObj.costFio = apis.fio.sufToAmount(fee_collected);
-        retObj.costUsdc =
-          (apis.fio.sufToAmount(fee_collected) * domainCostUsdc) /
-          domainCostFio;
-      }
-
-      regItems.push(retObj);
-
-      for (let i = updatedCart.length - 1; i >= 0; i--) {
-        if (updatedCart[i].id === fioName) {
-          updatedCart.splice(i, 1);
-        }
-      }
-    }
-  }
-
-  return { errItems, regItems, updatedCart };
-};
 
 export const priceToNumber = (price: string) => +parseFloat(price).toFixed(2);
